requirements.txt
fastapi 
uvicorn 
langchain 
langchain-community 
langchain-text-splitters
openai 
langchain-openai 
faiss-cpu
python-multipart
python-dotenv
supabase
psycopg2-binary
pgvector
gunicorn